反則回りの解消
