+++ Assets/Scripts/SoccerGame/SoccerGame.asmdef
@@ -5,7 +5,8 @@
         "GUID:26d67a36a1a29324282c9ec60123e88f",
         "GUID:560b04d1a97f54a4e82edc0cbbb69285",
         "GUID:f51ebe6a0ceec4240a699833d6309b23",
-        "GUID:6910bb43451d7f349bdc3a9b4e719aaf"
+        "GUID:6910bb43451d7f349bdc3a9b4e719aaf",
+        "GUID:06d20a26a64f0d0429e7a1831c3af445"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
