+++ Assets/Scripts/UI/UI.asmdef
@@ -5,7 +5,8 @@
         "GUID:560b04d1a97f54a4e82edc0cbbb69285",
         "GUID:6910bb43451d7f349bdc3a9b4e719aaf",
         "GUID:53decd2d2628df540b11f13017eb7496",
-        "GUID:941d7a5c210855c4994464255274c181"
+        "GUID:941d7a5c210855c4994464255274c181",
+        "GUID:6055be8ebefd69e48b49212b09b47b2f"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
